refactor: moved used types to common package
