+++ packages/daemon/src/db/index.ts
@@ -6,12 +6,9 @@
  */
 import mysql, { Connection as MysqlConnection, Pool } from 'mysql2/promise';
 import {
-  TokenBalanceMap,
   DbTxOutput,
   StringMap,
   Wallet,
-  TxInput,
-  TxOutputWithIndex,
   EventTxInput,
   GenerateAddresses,
   AddressIndexMap,
@@ -23,6 +20,11 @@ import {
   Miner,
   TokenSymbolsRow,
 } from '../types';
+import {
+  TxInput,
+  TokenBalanceMap,
+  TxOutputWithIndex,
+} from '@wallet-service/common/src/types';
 import { isAuthority } from '../utils';
 import {
   AddressBalanceRow,
